% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{area}
\alias{area}
\alias{area.default}
\alias{area.polygon}
\title{Calculate Surface Area}
\usage{
area(x, ...)

\method{area}{default}(x, y, ...)

\method{area}{polygon}(p)
}
\description{
Functions to calculate surface area.
}
\section{Methods (by class)}{
\itemize{
\item \code{area(default)}: Calculate surface area of a polygon using its vertex coordinates.

\item \code{area(polygon)}: Calculate the area of a \code{polygon} object.

}}
\examples{
x <- c(0, 0, 0.5, 1, 1, 0)
y <- c(0, 1, 0.5, 1, 0, 0)
area(x, y) # Calculate area of polygon.
}
