% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox}
\alias{bbox}
\alias{bbox.default}
\alias{bbox.polygon}
\title{Calculate Bounding Box}
\usage{
bbox(x, ...)

\method{bbox}{default}(x, y, ...)

\method{bbox}{polygon}(p, ...)
}
\description{
Functions to calculate bounding box of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Calculate the bounding box coordinates for a set of coordinate points.

\item \code{polygon}: Calculate surface area of a polygon using its vertex coordinates.
}}

\examples{
x <- c(0, 0, 0.5, 1, 1, 0)
y <- c(0, 1, 0.5, 1, 0, 0)
bbox(x, y) # Calculate area of polygon.

}
