% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gboxplot.R
\name{dboxplot}
\alias{dboxplot}
\alias{excel}
\alias{excel.default}
\alias{excel.data.frame}
\title{Export data frame to Excel}
\usage{
dboxplot(
  y,
  x,
  labels,
  width = 0.5,
  xlab = "",
  ylab = "",
  alpha = 0.05,
  lwd = 1,
  col = "black",
  bg = "white",
  ...
)
}
\arguments{
\item{x}{Data frame.}

\item{header}{Logical value specifying whether data frame column names should be exported.}

\item{row.names}{Logical value specifying whether data frame row names should be exported.}

\item{Further}{arguments passed onto the \code{\link[utils]{write.table}} function.}
}
\description{
This function dumps a data frame into a temporary tab-delimited table and opens an Excel session for the file.
}
\examples{

# Send a vector to Excel:
excel(1:10)

# Send a 25 x 4 matrix of random numbers to Excel, with no headers:
x <- matrix(runif(100), ncol = 4)
excel(x, header = FALSE)

}
