% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R, R/draw.points.R
\name{draw.points}
\alias{draw.points}
\alias{draw.polygon}
\title{Draw Graphics Objects}
\usage{
draw.points(
  x = NULL,
  y = NULL,
  n = NULL,
  tol = 0.005,
  snap = FALSE,
  snap.to.grid = FALSE,
  snap.to.line = FALSE,
  grid.increment,
  grid.offset = 0,
  pch = 21,
  bg = "black",
  ...
)

draw.polygon(
  x = NULL,
  y = NULL,
  n,
  tol = 0.005,
  snap = TRUE,
  snap.to.grid = FALSE,
  grid.increment,
  grid.offset = 0,
  pch = 21,
  bg = "black",
  ...
)

draw.points(
  x = NULL,
  y = NULL,
  n = NULL,
  tol = 0.005,
  snap = FALSE,
  snap.to.grid = FALSE,
  snap.to.line = FALSE,
  grid.increment,
  grid.offset = 0,
  pch = 21,
  bg = "black",
  ...
)
}
\arguments{
\item{x, y}{Horizontal coordinates plotted prior to drawing.}

\item{snap}{Logical value specifying whether points are to be snapped to the nearest \code{(x,y)} coordinates.}

\item{snap.togrid}{Logical value specifying whether points are to be snapped to the reference grid.}
}
\description{
Draw Graphics Objects

Draw Graphics Objects
}
\section{Functions}{
\itemize{
\item \code{draw.points}: Draw a points interactively on a plot.

\item \code{draw.polygon}: Draw a polygon interactively on a plot.

\item \code{draw.points}: Draw a points interactively on a plot.
}}

