% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.bar.R
\name{error.bar}
\alias{error.bar}
\title{Draw Error Bars}
\usage{
error.bar(x, y, sigma, lower, upper, add = TRUE, ...)
}
\arguments{
\item{x}{Horizontal coordinates.}

\item{y}{Central value(s) or lower and upper bound(s).}

\item{sigma}{Error deviation from \code{y} to be used to define the lower and upper error bounds.}

\item{lower}{Lower error bound or deviation (if \code{sigma} is specified).}

\item{upper}{Upper error bound or deviation (if \code{sigma} is specified).}

\item{add}{Logical value specifying whether the error bars are to be added to the current plot (Default
is \code{TRUE}).}
}
\description{
Draw error bars with specified error levels at the specified coordinates.
}
\examples{
# Constant error:
error.bar(1:10, y = rnorm(10), sigma = 2, add = FALSE)
 
# Variable error:
error.bar(1:10, y = rnorm(10), sigma = 0.5 + runif(10), add = FALSE)

# Explicit bounds:
error.bar(1:10, lower = 2, upper = 2 + runif(10), add = FALSE)

}
