% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbarplot.R
\name{gbarplot}
\alias{gbarplot}
\title{Gulf Bar Plot}
\usage{
gbarplot(
  y,
  x = NULL,
  labels = NULL,
  width = 0.8,
  col = NULL,
  border = par("fg"),
  add = FALSE,
  yaxs = ifelse(all(y[!is.na(y)] <= 0) | all(y[!is.na(y)] >= 0), "i", "r"),
  legend = TRUE,
  ...
)
}
\arguments{
\item{y}{A numeric vector, table, matrix or data frame object containing numeric values which are
to be displayed as a bar plot. Each row of the matrix will be plotted as a separate bar.}

\item{x}{A vector specifying the numeric values or labels to be used for separating each bar along
the horizontal axis.}

\item{labels}{A character vector specifying labels to be used for labelling each bar along the horizontal
axis. The number of labels must equal or be consistent with those of \code{x} or \code{y}.}

\item{width}{A numeric value from 0 to 1 specifying the relative width of each bar. A value of zero corresponds
to a bar with no width while a value of one implies that adjacent bars are touching (i.e. there
is no space between the bars). The default value is \code{0.8}.}

\item{col}{Vector specifying the colour(s) of the bars to be drawn. Each column of 'y' is drawn in a separate
colour. The default colour for a single variable is "grey". If there is more than one variable
and \code{col} is unspecified, then colours are interpolated from a grayscale palette.}

\item{border}{Border colour(s) of the bars.}

\item{add}{A logical value specifying if bars should be added to an existing plot. The default is \code{FALSE}.}

\item{yaxs}{A character value specifying how to set the limits of the Y axis, see \code{\link[graphics]{par}}.}

\item{legend}{A logical value specifying whether a legend is to be drawn. The default is \code{TRUE}.
The legend entries are taken from the columns names of \code{y}.}

\item{\dots}{Further arguments to be passed onto the \code{\link[graphics]{plot}} function.}
}
\description{
Doug's alternative version of the \code{\link[graphics]{barplot}}
function. Numeric \code{x} values may be passed directly for proper positioning along the
X axis, allowing for easy overlay in existing plots. If both positive and negative 'y'
values are present, then the resulting stacked plot has bars separated above and below the X
axis.
}
\examples{

# Generate random vector:
y <- 10*rnorm(5)+5

# Simple barplot:
gbarplot(y)

# Same barplot with explicit 'x':
gbarplot(y, x = c(1, 2, 4, 5, 7))

# Add explicit labels:
gbarplot(y, x = c(1, 2, 4, 5, 7), labels = letters[1:5])

# Display bar plot:
y <- data.frame("Variable 1" = c(1, -2, NA, -4, 5), "Variable 2" = 1:5)
gbarplot(y, xlab = "Order", ylab = "y",
         labels = c("First", "Second", "Third", "Fourth", "Fifth"))

# Bar plot of lobster length-frequencies by sex:
x <- read.gulf(year = 2000, species = 2550, card = "len")
f <- t(freq(x, by = "sex"))
f <- as.data.frame(f[2:dim(f)[1], 1:2])
f[, 1] <- -f[, 1]  # Reverse sign of male frequencies.
names(f) <- sex.str(1:2) # Assign sex labels.
gbarplot(f)

# Display a frequency table overlayed by the true distribution:
y <- table(round(rnorm(250, sd = 5)))
gbarplot(y, col = "grey")
x <- seq(-15, 15, len = 100)
lines(x, 250 * dnorm(x, sd = 5), col = "red", lwd = 2)

}
