% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gboxplot.R
\name{gboxplot}
\alias{gboxplot}
\title{Boxplot}
\usage{
gboxplot(
  y,
  x,
  labels,
  width = 0.5,
  xlab = "",
  ylab = "",
  alpha = 0.05,
  lwd = 1,
  col = "black",
  bg = "white",
  ...
)
}
\description{
Draws a boxplot based on input data.
}
