% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdevice.R
\name{gdevice}
\alias{gdevice}
\title{Prepare Graphics Device}
\usage{
gdevice(type = "", file, width = 8.5, height = 8.5, layout, ...)
}
\arguments{
\item{type}{Character string specifying the type of graphics device to be created (e.g. 'pdf', 'jpg', ...).}

\item{file}{File name, excluding file extension suffix.}

\item{width, height}{Dimensions in inches.}

\item{layout}{Specify plot layout of graphics device.}
}
\description{
A wrapper function to prepare a graphics device for output.
}
\examples{
gdevice("pdf", "Test.pdf")
gdevice() # Equivalent to 'dev.new()'.
plot(runif(100), runif(100))
dev.off()
}
