% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{hline}
\alias{hline}
\alias{vline}
\title{Draw Horizontal or Vertical Lines}
\usage{
hline(x, lower, upper, label, at, ...)

vline(x, lower, upper, label, at, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{lower}{A vector which defines the line(s)' lower plot range.}

\item{upper}{A vector which defines the line(s)' upper plot range.}

\item{label}{Line label(s).}

\item{at}{Numeric value(s) specifying where along the line the labels are to be displayed.}

\item{...}{Arguments passed onto \code{\link[graphics]{lines}} and \code{\link[graphics]{text}}.}
}
\description{
These functions are used to draw horizontal or vertical lines on an existing plot.
}
\section{Functions}{
\itemize{
\item \code{hline}: Draw horizontal lines.

\item \code{vline}: Draw vertical lines.
}}

\examples{
  plot(c(0, 10), c(0, 10), type = "n")
  hline(0, col = "red", lty = "solid")
  vline(9, col = "red", lty = "dashed")
  vline(1:5, col = "black", lty = "solid", lwd = 2)
}
