% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.esonar.R
\name{plot.esonar}
\alias{plot.esonar}
\title{Plot \strong{eSonar} Data}
\usage{
\method{plot}{esonar}(x, ...)
}
\arguments{
\item{x}{\code{esonar} data object.}
}
\description{
Graphically display \code{esonar} data.
}
\examples{
plot.esonar(year = 2020, tow.id = "GP354F")

x <- read.esonar(year = 2020, tow.id = "GP354F")
plot(x)

}
