% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scsset.R
\name{plot.scsset}
\alias{plot.scsset}
\title{Plot Snow Crab Trawl Data}
\usage{
\method{plot}{scsset}(x, tow.id, year, ...)
}
\arguments{
\item{x}{\code{scsset} data object.}

\item{tow.id}{Character string specifying a tow identifier.}

\item{year}{Survey year}
}
\description{
Displays auxiliary trawl data from the snow crab trawl survey.
}
\examples{
plot.scsset(year = 2020, tow.id = "GP354F")

x <- read.scsset(year = 2020)
plot(x[1, ])
}
