% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon}
\alias{polygon}
\alias{polygon.default}
\alias{as.polygon}
\alias{plot.polygon}
\alias{in.polygon}
\alias{which.polygon}
\alias{as.data.frame.polygon}
\alias{is.polygon}
\title{Polygon Functions}
\usage{
polygon(x, ...)

\method{polygon}{default}(x, ...)

as.polygon(x, y = NULL, hole = NULL, ...)

\method{plot}{polygon}(p, labels = FALSE, as.lines = FALSE, ...)

in.polygon(p, x, y, ...)

which.polygon(p, x, y, as.list = FALSE)

\method{as.data.frame}{polygon}(x)

is.polygon(p)
}
\arguments{
\item{x}{May be either a \sQuote{polygon} object, a list containing polygon
vertex definitions (with \sQuote{x} and \sQuote{y} coordinates or longitude
and latitude) or a numeric vector of horizontal polygon vertex coordinates.}

\item{y}{A numeric vector of vertical polygon vertex coordinates.}

\item{hole}{A logical vector specifying whether component polygons are to be
treated as holes. The length of the vector may be either the same as numeric
\sQuote{x} and \sQuote{y} or its length may be that of the number of
component polygons of the object. Polygons labelled as holes have negative
areas.}

\item{p}{List vector of polygon object definitions. Each element in the list
corresponds to a single polygon object. They contain two fields, \code{x}
and \code{y}, defining the vertices of each polygon.}
}
\value{
logical value \code{TRUE} is returned when a point (\code{x}, \code{y})
lies within and \code{FALSE} if it lies outside each polygon.
}
\description{
Set of functions to create, manipulate and query polygon objects.
}
\section{Functions}{

\describe{
  \item{\code{polygon}, \code{as.polygon}}{Creates or converts to a \sQuote{polygon} object.}
  \item{\code{plot.polygon}}{Graphically display a polygon object.}
  \item{\code{in.polygon}}{Determine whether a coordinate point lies within a polygon. A logical (\code{n} by \code{k}) matrix is returned where \code{n} is the number of
                           component polygons in \code{p} and \code{k} is the length of the \code{x}
                           and \code{y} vectors.}
  \item{\code{which.polygon}}{Returns the indices specifying to which polygon(s) in a list a point belongs.}
  \item{\code{area.polygon}}{Calculates the area of a polygon.}
  \item{\code{as.data.frame.polygon}}{Convert polygon object to data frame.}
  \item{\code{bbox.polygon}}{Determine bounding box for a polygon object.}
  \item{\code{draw.polygon}}{Draw a polygon on a plot.}
  \item{\code{is.polygon}}{Checks whether object is a \sQuote{polygon} object.}
}
}

\examples{
# Define set of polygon vertices:
x <- c(0, 0, 0.5, 1, 1, 0)
y <- c(0, 1, 0.5, 1, 0, 0)
p <- as.polygon(x, y)
plot(p) # Plot polygon.
area(p) # Calculate area of polygon.

# Generate set of random points:
x <- runif(1000)*2-0.5
y <- runif(1000)*2-0.5

# Plot points:
plot(c(-0.5, 1.5), c(-0.5, 1.5), type = "n")
index <- in.polygon(p, x, y)
points(x[index], y[index], pch = 21, bg = "red")
points(x[!index], y[!index], pch = 21, bg = "blue")

}
\seealso{
\code{\link[sp]{point.in.polygon}}
}
