% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.bar.R
\name{scale.bar}
\alias{scale.bar}
\title{Draw Scale Bar}
\usage{
\method{scale}{bar}(
  longitude,
  latitude,
  length = 100,
  units = "km",
  convert = TRUE,
  number = 4,
  double = TRUE,
  thickness = 0.05,
  pos = 1,
  margin = 0.25,
  cex = 0.6,
  colours = "black",
  ...
)
}
\arguments{
\item{longitude}{A numerical value specifying the longitude where the scale bar is to be drawn.}

\item{latitude}{A numerical value specifying the latitude where the scale bar is to be drawn.}

\item{length}{A numerical value specifying the length of the scale bar to be drawn. The units are specified by the \code{units} argument.}

\item{units}{A character string specifying the length units to be used. At present, the only options are \sQuote{km} or \sQuote{kilometers}.}

\item{convert}{A logical value specifying whether or not to convert the \code{length} argument to the degree scale.}

\item{number}{A numerical value specifying the number of intervals to include on the scale bar.}

\item{double}{A logical value specifying whether the scale bar is to be composed of a double or single bar.}

\item{thickness}{A numerical value specifying the thickness of the scale bar in inches.}

\item{pos}{A numerical value specifying the position of the scale bar. It may take the following values: 1 = Lower-left corner, 
2 = Lower-right corner, 3 = Upper-right corner and 4 = Upper-left corner.}

\item{margin}{A numerical value specifying the margin of separation between scale bar and the axis borders in inches. 
It is used in conjunction with the \code{pos} argument.}

\item{cex}{Character expansion value for the scale bar labels.}

\item{colours}{Scale bar colours.}
}
\description{
Display a scale bar on a map.
}
\examples{
map.new() # Create new window.
map("coast")    # Plot empty map.
scale.bar()

}
