% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scsset.R
\name{plot.scsset}
\alias{plot.scsset}
\title{Plot Snow Crab Trawl Data}
\usage{
\method{plot}{scsset}(
  x,
  tow.id,
  year,
  pdf = FALSE,
  path = getwd(),
  temperature = FALSE,
  headline = TRUE,
  wingspread = TRUE,
  tilt = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{scsset} data object.}

\item{tow.id}{Character string specifying a tow identifier.}

\item{year}{Survey year.}

\item{pdf}{Logical values specifying whether to output the figure to a pdf file.}

\item{path}{Path to which output file is to written.}

\item{temperature}{Logical values specifying whether to include temperature profile.}

\item{headline}{Logical values specifying whether to include trawl headline height measurements.}

\item{wingspread}{Logical values specifying whether to include trawl wing spread measurements.}

\item{tilt}{Logical values specifying whether to include tilt probe angles.}
}
\description{
Displays auxiliary trawl data from the snow crab trawl survey.
}
\examples{
plot.scsset(year = 2020, tow.id = "GP354F")

x <- read.scsset(year = 2020)
plot(x[1, ])
}
